Fix stat size on Linux and cleanup code (#95)
